Test: separate xGGQRCS tests from helper code
